refactor :: SMS 로직 이동
