+++ src/main/kotlin/hs/kr/equus/user/domain/admin/service/AdminTokenRefreshService.kt
@@ -1,7 +1,7 @@
 package hs.kr.equus.user.domain.admin.service
 
 import hs.kr.equus.user.global.security.jwt.JwtTokenProvider
-import hs.kr.equus.user.global.utils.token.dto.TokenResponse
+import hs.kr.equus.user.global.security.jwt.token.dto.TokenResponse
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
