+++ src/main/kotlin/hs/kr/equus/user/domain/user/service/UserTokenRefreshService.kt
@@ -1,7 +1,7 @@
 package hs.kr.equus.user.domain.user.service
 
 import hs.kr.equus.user.global.security.jwt.JwtTokenProvider
-import hs.kr.equus.user.global.utils.token.dto.TokenResponse
+import hs.kr.equus.user.global.security.jwt.token.dto.TokenResponse
 import org.springframework.stereotype.Service
 
 @Service
