+++ src/main/kotlin/hs/kr/equus/user/global/security/jwt/JwtTokenProvider.kt
@@ -9,7 +9,7 @@ import hs.kr.equus.user.global.exception.InvalidTokenException
 import hs.kr.equus.user.global.property.JwtProperties
 import hs.kr.equus.user.global.security.auth.AdminDetailsService
 import hs.kr.equus.user.global.security.auth.AuthDetailsService
-import hs.kr.equus.user.global.utils.token.dto.TokenResponse
+import hs.kr.equus.user.global.security.jwt.token.dto.TokenResponse
 import io.jsonwebtoken.*
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
 import org.springframework.security.core.Authentication
